Extract class to its own file
